Se crean los seeder, actualizan las migraciones y se empieza el crud con  Personal

+++ database/migrations/2019_08_18_020247_create_bajas_productos_table.php
@@ -19,7 +19,7 @@ class CreateBajasProductosTable extends Migration
             $table->date('fecha');
             $table->decimal('cantidad')->default(0);
             $table->integer('precio')->default(0);
-            $table->integer('id_producto');
+            $table->bigInteger('id_producto');
            
             $table->foreign('id_producto')->references('id_producto')->on('productos')->onDelete('cascade')->onUpdate('cascade');
             $table->timestamps();
